+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/DefaultLoadControl.java
@@ -24,6 +24,7 @@ import static java.lang.Math.min;
 
 import android.content.ContentResolver;
 import android.text.TextUtils;
+import androidx.annotation.GuardedBy;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import androidx.media3.common.C;
@@ -39,11 +40,16 @@ import androidx.media3.exoplayer.analytics.PlayerId;
 import androidx.media3.exoplayer.source.MediaSource.MediaPeriodId;
 import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
+import androidx.media3.exoplayer.upstream.Allocation;
 import androidx.media3.exoplayer.upstream.Allocator;
 import androidx.media3.exoplayer.upstream.DefaultAllocator;
+import androidx.media3.exoplayer.upstream.PlayerIdAwareAllocator;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /** The default {@link LoadControl} implementation. */
 @UnstableApi
@@ -171,6 +177,8 @@ public class DefaultLoadControl implements LoadControl {
   /** Builder for {@link DefaultLoadControl}. */
   public static final class Builder {
 
+    private final HashMap<String, Integer> playerTargetBufferBytes;
+
     @Nullable private DefaultAllocator allocator;
     private int minBufferMs;
     private int minBufferForLocalPlaybackMs;
@@ -189,6 +197,7 @@ public class DefaultLoadControl implements LoadControl {
 
     /** Constructs a new instance. */
     public Builder() {
+      playerTargetBufferBytes = new HashMap<>();
       minBufferMs = DEFAULT_MIN_BUFFER_MS;
       minBufferForLocalPlaybackMs = DEFAULT_MIN_BUFFER_FOR_LOCAL_PLAYBACK_MS;
       maxBufferMs = DEFAULT_MAX_BUFFER_MS;
@@ -351,10 +360,11 @@ public class DefaultLoadControl implements LoadControl {
     }
 
     /**
-     * Sets the target buffer size in bytes for each player. The actual overall target buffer size
-     * is this value multiplied by the number of players that use the load control simultaneously.
-     * If set to {@link C#LENGTH_UNSET}, the target buffer size of a player will be calculated based
-     * on the selected tracks of the player.
+     * Sets the target buffer size in bytes for each player. If set to {@link C#LENGTH_UNSET}, the
+     * target buffer size of a player will be calculated based on the selected tracks of the player.
+     *
+     * <p>This value will be ignored for the players with a {@code playerName} that has target
+     * buffer size set via {@link #setPlayerTargetBufferBytes(String, int)}.
      *
      * @param targetBufferBytes The target buffer size in bytes.
      * @return This builder, for convenience.
@@ -364,6 +374,26 @@ public class DefaultLoadControl implements LoadControl {
     public Builder setTargetBufferBytes(int targetBufferBytes) {
       checkState(!buildCalled);
       this.targetBufferBytes = targetBufferBytes;
+      return this;
+    }
+
+    /**
+     * Sets the target buffer size in bytes for a player with the specified {@code playerName}. When
+     * not set or set to {@link C#LENGTH_UNSET}, the target buffer size of a player will be the
+     * value set via {@link #setTargetBufferBytes(int)} if it is not {@link C#LENGTH_UNSET},
+     * otherwise it will be calculated based on the selected tracks of the player.
+     *
+     * @param playerName The name of the player. The same name must be set to the player via {@link
+     *     ExoPlayer.Builder#setName(String)} in order to be effective at the created {@link
+     *     DefaultLoadControl}.
+     * @param playerTargetBufferBytes The target buffer size in bytes for the specified player.
+     * @return This builder, for convenience.
+     * @throws IllegalStateException If {@link #build()} has already been called.
+     */
+    @CanIgnoreReturnValue
+    public Builder setPlayerTargetBufferBytes(String playerName, int playerTargetBufferBytes) {
+      checkState(!buildCalled);
+      this.playerTargetBufferBytes.put(playerName, playerTargetBufferBytes);
       return this;
     }
 
@@ -468,7 +498,8 @@ public class DefaultLoadControl implements LoadControl {
           prioritizeTimeOverSizeThresholds,
           prioritizeTimeOverSizeThresholdsForLocalPlayback,
           backBufferDurationMs,
-          retainBackBufferFromKeyframe);
+          retainBackBufferFromKeyframe,
+          playerTargetBufferBytes);
     }
   }
 
@@ -488,7 +519,8 @@ public class DefaultLoadControl implements LoadControl {
   private final boolean prioritizeTimeOverSizeThresholdsForLocalPlayback;
   private final long backBufferDurationUs;
   private final boolean retainBackBufferFromKeyframe;
-  private final HashMap<PlayerId, PlayerLoadingState> loadingStates;
+  private final ImmutableMap<String, Integer> playerTargetBufferBytesOverwrites;
+  private final ConcurrentHashMap<PlayerId, PlayerLoadingState> loadingStates;
 
   private long threadId;
 
@@ -525,7 +557,8 @@ public class DefaultLoadControl implements LoadControl {
       boolean prioritizeTimeOverSizeThresholds,
       boolean prioritizeTimeOverSizeThresholdsForLocalPlayback,
       int backBufferDurationMs,
-      boolean retainBackBufferFromKeyframe) {
+      boolean retainBackBufferFromKeyframe,
+      Map<String, Integer> playerTargetBufferBytes) {
     assertGreaterOrEqual(bufferForPlaybackMs, 0, "bufferForPlaybackMs", "0");
     assertGreaterOrEqual(
         bufferForPlaybackForLocalPlaybackMs, 0, "bufferForPlaybackForLocalPlaybackMs", "0");
@@ -578,8 +611,42 @@ public class DefaultLoadControl implements LoadControl {
         prioritizeTimeOverSizeThresholdsForLocalPlayback;
     this.backBufferDurationUs = msToUs(backBufferDurationMs);
     this.retainBackBufferFromKeyframe = retainBackBufferFromKeyframe;
-    loadingStates = new HashMap<>();
+    loadingStates = new ConcurrentHashMap<>();
+    playerTargetBufferBytesOverwrites = ImmutableMap.copyOf(playerTargetBufferBytes);
     threadId = C.INDEX_UNSET;
+  }
+
+  protected DefaultLoadControl(
+      DefaultAllocator allocator,
+      int minBufferMs,
+      int minBufferForLocalPlaybackMs,
+      int maxBufferMs,
+      int maxBufferForLocalPlaybackMs,
+      int bufferForPlaybackMs,
+      int bufferForPlaybackForLocalPlaybackMs,
+      int bufferForPlaybackAfterRebufferMs,
+      int bufferForPlaybackAfterRebufferForLocalPlaybackMs,
+      int targetBufferBytes,
+      boolean prioritizeTimeOverSizeThresholds,
+      boolean prioritizeTimeOverSizeThresholdsForLocalPlayback,
+      int backBufferDurationMs,
+      boolean retainBackBufferFromKeyframe) {
+    this(
+        allocator,
+        minBufferMs,
+        minBufferForLocalPlaybackMs,
+        maxBufferMs,
+        maxBufferForLocalPlaybackMs,
+        bufferForPlaybackMs,
+        bufferForPlaybackForLocalPlaybackMs,
+        bufferForPlaybackAfterRebufferMs,
+        bufferForPlaybackAfterRebufferForLocalPlaybackMs,
+        targetBufferBytes,
+        prioritizeTimeOverSizeThresholds,
+        prioritizeTimeOverSizeThresholdsForLocalPlayback,
+        backBufferDurationMs,
+        retainBackBufferFromKeyframe,
+        ImmutableMap.of());
   }
 
   @Override
@@ -601,6 +668,7 @@ public class DefaultLoadControl implements LoadControl {
       LoadControl.Parameters parameters,
       TrackGroupArray trackGroups,
       @NullableType ExoTrackSelection[] trackSelections) {
+    int targetBufferBytesOverwrite = getTargetBufferBytesOverwrite(parameters.playerId);
     checkNotNull(loadingStates.get(parameters.playerId)).targetBufferBytes =
         targetBufferBytesOverwrite == C.LENGTH_UNSET
             ? calculateTargetBufferBytes(parameters, trackSelections)
@@ -622,8 +690,8 @@ public class DefaultLoadControl implements LoadControl {
   }
 
   @Override
-  public Allocator getAllocator() {
-    return allocator;
+  public Allocator getAllocator(PlayerId playerId) {
+    return new PlayerIdFilteringAllocatorImpl(playerId);
   }
 
   @Override
@@ -641,7 +709,8 @@ public class DefaultLoadControl implements LoadControl {
     PlayerLoadingState playerLoadingState = checkNotNull(loadingStates.get(parameters.playerId));
     boolean isLocalPlayback = isLocalPlayback(parameters);
     boolean targetBufferSizeReached =
-        allocator.getTotalBytesAllocated() >= calculateTotalTargetBufferBytes();
+        getTotalBufferBytesAllocated(parameters.playerId)
+            >= getTargetBufferBytes(parameters.playerId);
     long minBufferUs = getMinBufferUs(isLocalPlayback);
     long maxBufferUs = getMaxBufferUs(isLocalPlayback);
     if (parameters.playbackSpeed > 1) {
@@ -683,12 +752,13 @@ public class DefaultLoadControl implements LoadControl {
     return minBufferDurationUs <= 0
         || bufferedDurationUs >= minBufferDurationUs
         || (!prioritizeTimeOverSizeThresholds(isLocalPlayback)
-            && allocator.getTotalBytesAllocated() >= calculateTotalTargetBufferBytes());
+            && getTotalBufferBytesAllocated(parameters.playerId)
+                >= getTargetBufferBytes(parameters.playerId));
   }
 
   @Override
   public boolean shouldContinuePreloading(
-      Timeline timeline, MediaPeriodId mediaPeriodId, long bufferedDurationUs) {
+      PlayerId playerId, Timeline timeline, MediaPeriodId mediaPeriodId, long bufferedDurationUs) {
     for (PlayerLoadingState playerLoadingState : loadingStates.values()) {
       if (playerLoadingState.isLoading) {
         return false;
@@ -744,11 +814,21 @@ public class DefaultLoadControl implements LoadControl {
 
   private void resetPlayerLoadingState(PlayerId playerId) {
     PlayerLoadingState playerLoadingState = checkNotNull(loadingStates.get(playerId));
+    int targetBufferBytesOverwrite = getTargetBufferBytesOverwrite(playerId);
     playerLoadingState.targetBufferBytes =
-        targetBufferBytesOverwrite == C.LENGTH_UNSET
-            ? DEFAULT_MIN_BUFFER_SIZE
-            : targetBufferBytesOverwrite;
+        targetBufferBytesOverwrite != C.LENGTH_UNSET
+            ? targetBufferBytesOverwrite
+            : DEFAULT_MIN_BUFFER_SIZE;
     playerLoadingState.isLoading = false;
+  }
+
+  private int getTargetBufferBytesOverwrite(PlayerId playerId) {
+    Integer playerTargetBufferBytesOverwrite = playerTargetBufferBytesOverwrites.get(playerId.name);
+    if (playerTargetBufferBytesOverwrite != null
+        && playerTargetBufferBytesOverwrite != C.LENGTH_UNSET) {
+      return playerTargetBufferBytesOverwrite;
+    }
+    return targetBufferBytesOverwrite;
   }
 
   private void removePlayer(PlayerId playerId) {
@@ -831,8 +911,101 @@ public class DefaultLoadControl implements LoadControl {
     checkArgument(value1 >= value2, "%s cannot be less than %s", name1, name2);
   }
 
+  private int getTotalBufferBytesAllocated(PlayerId playerId) {
+    return checkNotNull(loadingStates.get(playerId)).getAllocatedCounts()
+        * allocator.getIndividualAllocationLength();
+  }
+
+  private int getTargetBufferBytes(PlayerId playerId) {
+    return checkNotNull(loadingStates.get(playerId)).targetBufferBytes;
+  }
+
   private static class PlayerLoadingState {
     public boolean isLoading;
     public int targetBufferBytes;
+
+    @GuardedBy("this")
+    private int allocatedCounts;
+
+    public synchronized void increaseAllocatedCounts() {
+      allocatedCounts++;
+    }
+
+    public synchronized void decreaseAllocatedCounts() {
+      allocatedCounts--;
+    }
+
+    public synchronized int getAllocatedCounts() {
+      return allocatedCounts;
+    }
+  }
+
+  private final class PlayerIdFilteringAllocatorImpl implements PlayerIdAwareAllocator {
+    @GuardedBy("this")
+    private final HashMap<Allocation, PlayerId> allocationPlayerIdMap;
+
+    @GuardedBy("this")
+    private PlayerId playerId;
+
+    public PlayerIdFilteringAllocatorImpl(PlayerId playerId) {
+      this.allocationPlayerIdMap = new HashMap<>();
+      this.playerId = playerId;
+    }
+
+    @Override
+    public synchronized void setPlayerId(PlayerId playerId) {
+      this.playerId = playerId;
+    }
+
+    @Override
+    public Allocation allocate() {
+      Allocation allocation = allocator.allocate();
+      synchronized (this) {
+        allocationPlayerIdMap.put(allocation, playerId);
+        @Nullable PlayerLoadingState playerLoadingState = loadingStates.get(playerId);
+        if (playerLoadingState != null) {
+          playerLoadingState.increaseAllocatedCounts();
+        }
+      }
+      return allocation;
+    }
+
+    @Override
+    public void release(Allocation allocation) {
+      allocator.release(allocation);
+      releaseInternal(allocation);
+    }
+
+    @Override
+    public void release(@Nullable AllocationNode allocationNode) {
+      allocator.release(allocationNode);
+      while (allocationNode != null) {
+        releaseInternal(allocationNode.getAllocation());
+        allocationNode = allocationNode.next();
+      }
+    }
+
+    @Override
+    public void trim() {
+      allocator.trim();
+    }
+
+    @Override
+    public synchronized int getTotalBytesAllocated() {
+      return getTotalBufferBytesAllocated(playerId);
+    }
+
+    @Override
+    public int getIndividualAllocationLength() {
+      return allocator.getIndividualAllocationLength();
+    }
+
+    private synchronized void releaseInternal(Allocation allocation) {
+      PlayerId playerId = checkNotNull(allocationPlayerIdMap.remove(allocation));
+      @Nullable PlayerLoadingState playerLoadingState = loadingStates.get(playerId);
+      if (playerLoadingState != null) {
+        playerLoadingState.decreaseAllocatedCounts();
+      }
+    }
   }
 }