+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/preload/DefaultPreloadManager.java
@@ -43,6 +43,7 @@ import androidx.media3.exoplayer.LoadControl;
 import androidx.media3.exoplayer.PlaybackLooperProvider;
 import androidx.media3.exoplayer.RendererCapabilitiesList;
 import androidx.media3.exoplayer.RenderersFactory;
+import androidx.media3.exoplayer.analytics.PlayerId;
 import androidx.media3.exoplayer.source.DefaultMediaSourceFactory;
 import androidx.media3.exoplayer.source.MediaSource;
 import androidx.media3.exoplayer.source.SampleQueue;
@@ -498,7 +499,7 @@ public final class DefaultPreloadManager
             trackSelector,
             bandwidthMeter,
             rendererCapabilitiesList.getRendererCapabilities(),
-            builder.loadControlSupplier.get().getAllocator(),
+            builder.loadControlSupplier.get().getAllocator(PlayerId.UNSET),
             preloadLooper);
     @Nullable Cache cache = builder.cache;
     if (cache != null) {